requirements.txt
# Core
pytz==2022.7.1
Django==4.1.7
django-configurations==2.4
gunicorn==20.1.0

# For the persistence stores
psycopg2-binary==2.9.5

# Model Tools
django-model-utils==4.3.1
django_unique_upload==0.2.1

# Rest apis
djangorestframework==3.14.0

# Milvus
pymilvus==2.2.8

# Developer Tools
ipdb==0.13.13
ipython==8.11.0
mkdocs==1.4.3
flake8==6.0.0

# Testing
mock==5.0.1
factory-boy==3.2.1
coverage==7.2.1

# Static and Media Storage
django-storages==1.13.2
boto3==1.26.89

# ML stuff
jax==0.4.10
flax==0.6.10
librosa==0.10.0.post2
matplotlib
soundfile==0.12.1
matplotlib==3.7.1
tensorflow==2.13.0-rc0